3D OBJ Models
git-svn-id: https://mediacenter-pc/svn/GPGPU/CUDA/RaytracingEngine/trunk@557 c4de0f11-4f5c-764a-a4aa-d9e147179871

+++ RayTracingEngineStub.cpp
@@ -406,7 +406,8 @@ extern "C" RAYTRACINGENGINE_API int RayTracer_LoadMolecule(
       static_cast<GeometryType>(geometryType),
       static_cast<float>(defaultAtomSize), 
       static_cast<float>(defaultStickSize),
-      atomMaterialType, scale );
+      atomMaterialType, 
+      static_cast<float>(scale) );
    return gpuKernel->compactBoxes(true);
 }
